lasalle_dmre_backend-master (1)/lasalle_dmre_backend-master/requirements.txt
asgiref==3.5.2
certifi
charset-normalizer==2.1.1
Cython==0.29.28
distance-map==1.0.1
dj-database-url==1.0.0
Django==4.1.2
django-cors-headers==3.13.0
django-on-heroku==1.1.2
djangorestframework==3.14.0
gcloud==0.18.3
gensim==4.2.0
googleapis-common-protos==1.56.4
gunicorn==20.1.0
httplib2==0.20.4
idna==3.4
jws==0.1.3
llvmlite==0.39.1
numba==0.56.3
numpy==1.23.4
oauth2client==4.1.3
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
protobuf==4.21.8
psycopg2-binary==2.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.15.0
pyparsing==3.0.9
Pyrebase4==4.5.0
python-decouple==3.6
python-jwt==2.0.1
pytz==2022.5
requests==2.28.1
requests-toolbelt==0.10.0
rsa==4.9
scipy==1.9.3
six==1.16.0
smart-open==6.2.0
sqlparse==0.4.3
tzdata==2022.5
urllib3==1.26.12
whitenoise==6.2.0
wincertstore==0.2
